+++ packages/core/src/components/index.ts
@@ -28,6 +28,7 @@ export * from './graphs/skeleton';
 export * from './graphs/skeleton-lines';
 export * from './graphs/treemap';
 export * from './graphs/radar';
+export * from './graphs/wordcloud';
 
 // Layout
 export * from './layout/spacer';