+++ packages/core/src/configuration.ts
@@ -17,12 +17,14 @@ import {
 	RadarChartOptions,
 	ComboChartOptions,
 	TreemapChartOptions,
+	WorldCloudChartOptions,
 	// Components
 	GridOptions,
 	RulerOptions,
 	AxesOptions,
 	TimeScaleOptions,
 	TooltipOptions,
+	WordCloudChartTooltipOptions,
 	LegendOptions,
 	StackedBarOptions,
 	MeterChartOptions,
@@ -36,6 +38,7 @@ import {
 	TruncationTypes,
 	ToolbarControlTypes,
 	ZoomBarTypes,
+	LegendItemType,
 } from './interfaces';
 import enUSLocaleObject from 'date-fns/locale/en-US/index';
 
@@ -64,6 +67,7 @@ const legend: LegendOptions = {
 	truncation: standardTruncationOptions,
 	alignment: Alignments.LEFT,
 	order: null,
+	additionalItems: [],
 };
 
 /**
@@ -74,11 +78,13 @@ export const grid: GridOptions = {
 		// set enable to false will not draw grid and stroke of grid backdrop
 		enabled: true,
 		numberOfTicks: 15,
+		alignWithAxisTicks: false,
 	},
 	y: {
 		// set enable to false will not draw grid and stroke of grid backdrop
 		enabled: true,
 		numberOfTicks: 5,
+		alignWithAxisTicks: false,
 	},
 };
 
@@ -96,8 +102,8 @@ export const ruler: RulerOptions = {
 export const baseTooltip: TooltipOptions = {
 	enabled: true,
 	showTotal: true,
-	valueFormatter: (d) => d.toLocaleString(),
 	truncation: standardTruncationOptions,
+	groupLabel: 'Group',
 };
 
 // These options will be managed by Tools.mergeDefaultChartOptions
@@ -293,6 +299,7 @@ const stackedAreaChart = areaChart;
 const bubbleChart: BubbleChartOptions = Tools.merge({}, axisChart, {
 	bubble: {
 		radiusMapsTo: 'radius',
+		radiusLabel: 'Radius',
 		radiusRange: (chartSize, data) => {
 			const smallerChartDimension = Math.min(
 				chartSize.width,
@@ -309,9 +316,18 @@ const bubbleChart: BubbleChartOptions = Tools.merge({}, axisChart, {
 	points: {
 		filled: true,
 	},
+	legend: {
+		additionalItems: [
+			{
+				type: LegendItemType.RADIUS,
+				name: 'Radius',
+			},
+		],
+	},
 } as BubbleChartOptions);
 
 /**
+<<<<<<< HEAD
  * options specific to bullet charts
  */
 const bulletChart: BulletChartOptions = Tools.merge({}, axisChart, {
@@ -319,6 +335,30 @@ const bulletChart: BulletChartOptions = Tools.merge({}, axisChart, {
 		performanceAreaTitles: ['Poor', 'Satisfactory', 'Great'],
 	},
 } as BulletChartOptions);
+=======
+ * options specific to word cloud charts
+ */
+const wordCloudChart: WorldCloudChartOptions = Tools.merge({}, chart, {
+	tooltip: Tools.merge({}, baseTooltip, {
+		wordLabel: 'Word',
+		valueLabel: 'Value',
+	}) as WordCloudChartTooltipOptions,
+	wordCloud: {
+		fontSizeMapsTo: 'value',
+		fontSizeRange: (chartSize, data) => {
+			const smallerChartDimension = Math.min(
+				chartSize.width,
+				chartSize.height
+			);
+			return [
+				(smallerChartDimension * 20) / 400,
+				(smallerChartDimension * 75) / 400,
+			];
+		},
+		wordMapsTo: 'word',
+	},
+} as WorldCloudChartOptions);
+>>>>>>> c3c5593bc6209cb7d9dd49ae03e83015a9f417da
 
 /**
  * options specific to pie charts
@@ -447,6 +487,7 @@ export const options = {
 	gaugeChart,
 	comboChart,
 	treemapChart,
+	wordCloudChart,
 };
 
 export * from './configuration-non-customizable';