+++ packages/core/src/model.ts
@@ -338,8 +338,11 @@ export class ChartModel {
 	 * @param newOptions New options to be set
 	 */
 	setOptions(newOptions) {
+		const options = this.getOptions();
+		Tools.updateLegendAdditionalItems(options, newOptions);
+
 		this.set({
-			options: Tools.merge(this.getOptions(), newOptions),
+			options: Tools.merge(options, newOptions),
 		});
 	}
 
@@ -526,7 +529,7 @@ export class ChartModel {
 	 */
 	protected transformToTabularData(data) {
 		console.warn(
-			"We've updated the charting data format to be tabular by default. The current format you're using is deprecated and will be removed in v1.0, read more here https://carbon-design-system.github.io/carbon-charts/?path=/story/tutorials--tabular-data-format"
+			"We've updated the charting data format to be tabular by default. The current format you're using is deprecated and will be removed in v1.0, read more here https://carbon-design-system.github.io/carbon-charts/?path=/story/docs-tutorials--tabular-data-format"
 		);
 		const tabularData = [];
 		const { datasets, labels } = data;